删除节奏风暴相关代码
