+++ bilive/listener.ts
@@ -1,12 +1,9 @@
 import * as request from 'request'
 import * as tools from './lib/tools'
 import { EventEmitter } from 'events'
-import { roomInfo } from './online'
 import { AppClient } from './lib/app_client'
 import { CommentClient } from './lib/comment_client'
-import { SYS_MSG, SYS_GIFT } from './lib/danmaku.type'
-import { apiLiveOrigin, rafflePathname, lightenPathname, _options, smallTVPathname } from './index'
-import { BiliveClient, message, beatStormInfo, smallTVInfo, raffleInfo, lightenInfo, appLightenInfo, debugInfo } from './lib/bilive_client'
+import { liveOrigin, apiLiveOrigin, smallTVPathname, rafflePathname, lightenPathname, _options } from './index'
 /**
  * 监听服务器消息
  * 
@@ -27,14 +24,6 @@ export class Listener extends EventEmitter {
    */
   private _CommentClient: CommentClient
   /**
-   * 用于接收服务器消息
-   * 
-   * @private
-   * @type {BiliveClient}
-   * @memberof Listener
-   */
-  private _Client: BiliveClient
-  /**
    * 小电视ID
    * 
    * @private
@@ -42,14 +31,6 @@ export class Listener extends EventEmitter {
    * @memberof Listener
    */
   private _smallTVID: number = 0
-  /**
-   * 节奏风暴ID
-   * 
-   * @private
-   * @type {number}
-   * @memberof Listener
-   */
-  private _beatStormID: number = 0
   /**
    * 抽奖ID
    * 
@@ -89,19 +70,6 @@ export class Listener extends EventEmitter {
       .on('SYS_MSG', this._SYSMSGHandler.bind(this))
       .on('SYS_GIFT', this._SYSGiftHandler.bind(this))
       .Connect()
-    let apiOrigin = config.apiOrigin
-      , apiKey = config.apiKey
-    if (apiOrigin === '' || apiKey === '') return
-    this._Client = new BiliveClient(apiOrigin, apiKey)
-    this._Client
-      .on('serverError', (error) => { tools.Log('与监听服务器断开五分钟', error) })
-      .on('sysmsg', (message: message) => { tools.Log('系统消息:', message.msg) })
-      .on('smallTV', this._SmallTVHandler.bind(this))
-      .on('beatStorm', this._BeatStormHandler.bind(this))
-      .on('raffle', this._RaffleHandler.bind(this))
-      .on('lighten', this._LightenHandler.bind(this))
-      .on('debug', this._DebugHandler.bind(this))
-      .Connect()
   }
   /**
    * 监听弹幕系统消息
@@ -129,7 +97,7 @@ export class Listener extends EventEmitter {
       let url = apiLiveOrigin + rafflePathname
         , roomID = dataJson.real_roomid
       this._RaffleCheck(url, roomID, 'raffle')
-      this._AppLightenCheck(roomID)
+      // this._AppLightenCheck(roomID)
     }
     else if (dataJson.giftId === 84) {
       let roomID = dataJson.real_roomid
@@ -137,21 +105,18 @@ export class Listener extends EventEmitter {
           uri: `${apiLiveOrigin}${lightenPathname}/getLiveInfo?roomid=${roomID}`,
           json: true,
           headers: {
-            'Referer': `https://live.bilibili.com/${roomID}`
+            'Referer': `${liveOrigin}/${roomID}`
           }
         }
         , lightenCheck = await tools.XHR<lightenCheck>(check).catch(tools.Error)
       if (lightenCheck != null && lightenCheck.body.code === 0 && lightenCheck.body.data.length > 0) {
         lightenCheck.body.data.forEach(value => {
-          let message: message = {
+          let message: raffleMSG = {
             cmd: 'lighten',
-            data: {
-              roomID,
-              id: value.lightenId,
-              rawData: dataJson
-            }
+            roomID,
+            id: value.lightenId
           }
-          this._LightenHandler(message)
+          this._RaffleHandler(message)
         })
       }
     }
@@ -162,35 +127,41 @@ export class Listener extends EventEmitter {
    * @private
    * @param {string} url 
    * @param {number} roomID 
-   * @param {string} raffle 
+   * @param {('smallTV' | 'raffle' | 'lighten')} raffle 
    * @memberof Listener
    */
-  private async _RaffleCheck(url: string, roomID: number, raffle: string) {
+  private async _RaffleCheck(url: string, roomID: number, raffle: 'smallTV' | 'raffle' | 'lighten') {
     let check: request.Options = {
       uri: `${url}/check?roomid=${roomID}`,
       json: true,
       headers: {
-        'Referer': `https://live.bilibili.com/${roomID}`
+        'Referer': `${liveOrigin}/${roomID}`
       }
     }
       , raffleCheck = await tools.XHR<raffleCheck>(check)
     if (raffleCheck.response.statusCode === 200 && raffleCheck.body.code === 0 && raffleCheck.body.data.length > 0) {
-      raffleCheck.body.data.forEach(value => {
-        let message: message = {
+      raffleCheck.body.data.forEach(data => {
+        let message: raffleMSG = {
           cmd: raffle,
-          data: {
-            roomID,
-            id: value.raffleId
-          }
+          roomID,
+          id: data.raffleId
         }
-        if (raffle === 'smallTV') this._SmallTVHandler(message)
-        else if (raffle === 'raffle') this._RaffleHandler(message)
-        else if (raffle === 'lighten') this._LightenHandler(message)
+        this._RaffleHandler(message)
+        // 临时
+        if (raffle === 'raffle') {
+          let message: appLightenMSG = {
+            cmd: 'appLighten',
+            roomID,
+            id: data.raffleId,
+            type: 'openfire'
+          }
+          this._RaffleHandler(message)
+        }
       })
     }
   }
   /**
-   * 检查app房间信息
+   * 检查客户端房间信息
    * 
    * @private
    * @param {number} roomID 
@@ -206,149 +177,50 @@ export class Listener extends EventEmitter {
       roomInfo.body.data.event_corner.forEach(event => {
         let type = event.event_type.split('-')
         if (type.length !== 2) return
-        let message: message = {
+        let message: appLightenMSG = {
           cmd: 'appLighten',
-          data: {
-            roomID,
-            id: parseInt(type[1]),
-            type: type[0]
-          }
+          roomID,
+          id: parseInt(type[1]),
+          type: type[0]
         }
-        this._AppLightenHandler(message)
+        this._RaffleHandler(message)
       })
     }
-  }
-  /**
-   * 监听小电视消息
-   * 
-   * @private
-   * @param {message} message
-   * @memberof Listener
-   */
-  private _SmallTVHandler(message: message) {
-    let smallTVInfo = <smallTVInfo>message.data
-    if (this._smallTVID >= smallTVInfo.id) return
-    let roomID = smallTVInfo.roomID
-      , id = smallTVInfo.id
-    this._smallTVID = id
-    tools.Log(`房间 ${roomID} 赠送了第 ${id} 个小电视`)
-    this.emit('smallTV', smallTVInfo)
   }
   /**
    * 监听抽奖消息
    * 
    * @private
-   * @param {message} message
+   * @param {(raffleMSG | appLightenMSG)} raffleMSG 
    * @memberof Listener
    */
-  private _RaffleHandler(message: message) {
-    let raffleInfo = <raffleInfo>message.data
-    if (this._raffleID >= raffleInfo.id) return
-    let roomID = raffleInfo.roomID
-      , id = raffleInfo.id
-    this._raffleID = id
-    tools.Log(`房间 ${roomID} 赠送了第 ${id} 个活动道具`)
-    this.emit('raffle', raffleInfo)
+  private _RaffleHandler(raffleMSG: raffleMSG | appLightenMSG) {
+    let roomID = raffleMSG.roomID
+      , id = raffleMSG.id
+      , msg = ''
+    switch (raffleMSG.cmd) {
+      case 'smallTV':
+        if (this._smallTVID >= id) return
+        this._smallTVID = id
+        msg = '小电视'
+        break
+      case 'raffle':
+        if (this._raffleID >= id) return
+        this._raffleID = id
+        break
+      case 'lighten':
+        if (this._lightenID >= id) return
+        this._lightenID = id
+        break
+      case 'appLighten':
+        if (this._appLightenID >= id) return
+        this._appLightenID = id
+        msg = '客户端'
+        break
+      default:
+        return
+    }
+    this.emit('raffle', raffleMSG)
+    tools.Log(`房间 ${roomID} 开启了第 ${id} 轮${msg}抽奖`)
   }
-  /**
-   * 监听快速抽奖消息
-   * 
-   * @private
-   * @param {message} message
-   * @memberof Listener
-   */
-  private _LightenHandler(message: message) {
-    let lightenInfo = <lightenInfo>message.data
-    if (this._lightenID >= lightenInfo.id) return
-    let roomID = lightenInfo.roomID
-      , id = lightenInfo.id
-    this._lightenID = id
-    tools.Log(`房间 ${roomID} 赠送了第 ${id} 个活动道具`)
-    this.emit('lighten', lightenInfo)
-  }
-  /**
-   * 监听app快速抽奖消息
-   * 
-   * @private
-   * @param {message} message
-   * @memberof Listener
-   */
-  private _AppLightenHandler(message: message) {
-    let appLightenInfo = <appLightenInfo>message.data
-    if (this._appLightenID >= appLightenInfo.id) return
-    let roomID = appLightenInfo.roomID
-      , id = appLightenInfo.id
-    this._appLightenID = id
-    tools.Log(`房间 ${roomID} 赠送了第 ${id} 个活动道具`)
-    this.emit('appLighten', appLightenInfo)
-  }
-  /**
-   * 监听节奏风暴消息
-   * 
-   * @private
-   * @param {message} message
-   * @memberof Listener
-   */
-  private _BeatStormHandler(message: message) {
-    let beatStormInfo = <beatStormInfo>message.data
-    if (this._beatStormID >= beatStormInfo.id) return
-    let roomID = beatStormInfo.roomID
-      , id = beatStormInfo.id
-    this._beatStormID = id
-    tools.Log(`房间 ${roomID} 赠送了第 ${id} 个节奏风暴`)
-    this.emit('beatStorm', beatStormInfo)
-  }
-  /**
-   * 远程调试
-   * 
-   * @private
-   * @param {message} message 
-   * @memberof Listener
-   */
-  private _DebugHandler(message: message) {
-    let debugInfo = <debugInfo>message.data
-    tools.Log('远程调试信息:', debugInfo)
-    this.emit('debug', debugInfo)
-  }
-}
-/**
- * 抽奖检查
- * 
- * @export
- * @interface raffleCheck
- */
-interface raffleCheck {
-  code: number
-  msg: string
-  message: string
-  data: raffleCheckData[]
-}
-interface raffleCheckData {
-  raffleId: number
-  type: 'small_tv' | string
-  form: string
-  from_user: {
-    uname: string
-    face: string
-  }
-  time: number
-  status: number
-}
-/**
- * 快速抽奖检查
- * 
- * @export
- * @interface lightenCheck
- */
-interface lightenCheck {
-  code: number
-  msg: string
-  message: string
-  data: lightenCheckData[]
-}
-interface lightenCheckData {
-  type: string
-  lightenId: number
-  time: number
-  status: boolean
 }
