+++ bilive/online.ts
@@ -2,7 +2,7 @@ import * as request from 'request'
 import * as tools from './lib/tools'
 import { EventEmitter } from 'events'
 import { AppClient } from './lib/app_client'
-import { apiLiveOrigin, _options, cookieJar, userData } from './index'
+import { liveOrigin, apiLiveOrigin, _options, cookieJar } from './index'
 /**
  * 挂机得经验
  * 
@@ -46,7 +46,7 @@ export class Online extends EventEmitter {
           jar: cookieJar[uid],
           json: true,
           headers: {
-            'Referer': `https://live.bilibili.com/${roomID}`
+            'Referer': `${liveOrigin}/${roomID}`
           }
         }
           , heartPC = await tools.XHR<userOnlineHeart>(online)
@@ -171,7 +171,7 @@ export class Online extends EventEmitter {
         jar: cookieJar[uid],
         json: true,
         headers: {
-          'Referer': `https://live.bilibili.com/${roomIDs[0]}`
+          'Referer': `${liveOrigin}/${roomIDs[0]}`
         }
       }
       tools.XHR(task, 'WebView').catch(error => { tools.Error(userData.nickname, '日常活动', error) })
@@ -255,193 +255,4 @@ export class Online extends EventEmitter {
     }
     else tools.Log(userData.nickname, '应援团签到', '获取应援团列表失败', linkGroup.body)
   }
-}
-/**
- * 签到信息
- * 
- * @export
- * @interface signInfo
- */
-interface signInfo {
-  code: number
-  msg: string
-  data: signInfoData
-}
-interface signInfoData {
-  text: string
-  status: number
-  allDays: string
-  curMonth: string
-  newTask: number
-  hadSignDays: number
-  remindDays: number
-}
-/**
- * 在线心跳返回
- * 
- * @interface userOnlineHeart
- */
-interface userOnlineHeart {
-  code: number
-  msg: string
-}
-/**
- * 在线领瓜子宝箱
- * 
- * @interface currentTask
- */
-interface currentTask {
-  code: number
-  msg: string
-  data: currentTaskData
-}
-interface currentTaskData {
-  minute: number
-  silver: number
-  time_start: number
-  time_end: number
-}
-/**
- * 领瓜子答案提交返回
- * 
- * @interface award
- */
-interface award {
-  code: number
-  msg: string
-  data: awardData
-}
-interface awardData {
-  silver: number
-  awardSilver: number
-  isEnd: number
-}
-/**
- * 房间信息
- * 
- * @interface roomInfo
- */
-export interface roomInfo {
-  code: number
-  data: roomInfoData
-}
-interface roomInfoData {
-  room_id: number
-  mid: number
-  event_corner: roomInfoDataEvent[]
-}
-interface roomInfoDataEvent {
-  event_type: string
-  event_img: string
-}
-/**
- * 分享房间返回
- * 
- * @interface shareCallback
- */
-interface shareCallback {
-  code: number
-  msg: string
-  message: string
-}
-/**
- * 每日包裹
- * 
- * @interface getBagGift
- */
-interface getBagGift {
-  code: number
-}
-/**
- * 包裹信息
- * 
- * @interface bagInfo
- */
-interface bagInfo {
-  code: number
-  msg: string
-  message: string
-  data: bagInfoData[]
-}
-interface bagInfoData {
-  id: number
-  uid: number
-  gift_id: number
-  gift_num: number
-  expireat: number
-  gift_type: number
-  gift_name: string
-  gift_price: string
-  img: string
-  count_set: string
-  combo_num: number
-  super_num: number
-}
-/**
- * 赠送包裹礼物
- * 
- * @interface sendBag
- */
-interface sendBag {
-  code: number
-  msg: string
-  message: string
-  data: sendBagData
-}
-interface sendBagData {
-  tid: string
-  uid: number
-  uname: string
-  ruid: number
-  rcost: number
-  gift_id: number
-  gift_type: number
-  gift_name: string
-  gift_num: number
-  gift_action: string
-  gift_price: number
-  coin_type: string
-  total_coin: number
-  metadata: string
-  rnd: string
-}
-/**
- * 应援团
- * 
- * @interface linkGroup
- */
-interface linkGroup {
-  code: number
-  msg: string
-  message: string
-  data: linkGroupData
-}
-interface linkGroupData {
-  list: linkGroupInfo[]
-}
-interface linkGroupInfo {
-  group_id: number
-  owner_uid: number
-  owner_name: string
-  group_type: number
-  group_level: number
-  group_cover: string
-  group_name: string
-  group_notice: string
-  group_status: number
-}
-/**
- * 应援团签到返回
- * 
- * @interface signGroup
- */
-interface signGroup {
-  code: number
-  msg: string
-  message: string
-  data: signGroupData
-}
-interface signGroupData {
-  add_num: number
-  status: number
 }
